+++ lib/page/assets/index.dart
@@ -428,7 +428,6 @@ class _AssetsState extends State<Assets> {
     super.initState();
   }
 
-  // todo: add account index display
   @override
   Widget build(BuildContext context) {
     final Map dic = I18n.of(context).home;